requirements.txt
fastapi==0.108.0
pydantic==1.10.2
pymongo==4.6.1
python-decouple==3.8
uvicorn==0.25.0
anyio==3.6.2
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
